+++ RatRaceOnline/Models/ListViewItemModel.cs
@@ -1,0 +1,44 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RatRace3.Models
+{
+    public class ListViewItemModel : INotifyPropertyChanged
+    {
+        private string itemText;
+        private string itemValue;
+
+        public string ItemText
+        {
+            get { return itemText; }
+            set
+            {
+                itemText = value;
+                OnPropertyChanged("ItemText");
+            }
+        }
+
+        public string ItemValue
+        {
+            get { return itemValue; }
+            set
+            {
+                itemValue = value;
+                OnPropertyChanged("ItemValue");
+            }
+        }
+
+
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        public void OnPropertyChanged(string name)
+        {
+            if (this.PropertyChanged != null)
+                this.PropertyChanged(this, new PropertyChangedEventArgs(name));
+        }
+    }
+}
